requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
python-dotenv
openpyxl
python-multipart
bcrypt
